fix: resolve errors due to migration to angular 16
